+++ examples/advanced/src/gen/clients/axios/userService/createUser.ts
@@ -8,7 +8,7 @@ import type { CreateUserMutationRequest, CreateUserMutationResponse } from '../.
  * @link /user
  */
 export async function createUser(data?: CreateUserMutationRequest, config: Partial<RequestConfig<CreateUserMutationRequest>> = {}) {
-  const res = await client<CreateUserMutationResponse, unknown, CreateUserMutationRequest>({
+  const res = await client<CreateUserMutationResponse, Error, CreateUserMutationRequest>({
     method: 'post',
     url: '/user',
     baseURL: 'https://petstore3.swagger.io/api/v3',