+++ examples/client/src/gen/tagObject.ts
@@ -12,7 +12,7 @@ import type { RequestConfig } from '@kubb/plugin-client/client'
  * @link /store/inventory
  */
 export async function getInventory(config: Partial<RequestConfig> = {}) {
-  const res = await client<GetInventoryQueryResponse, unknown, unknown>({
+  const res = await client<GetInventoryQueryResponse, Error, unknown>({
     method: 'get',
     url: '/store/inventory',
     baseURL: 'https://petstore3.swagger.io/api/v3',