+++ examples/fetch/src/gen/updateUser.ts
@@ -12,7 +12,7 @@ export async function updateUser(
   data?: UpdateUserMutationRequest,
   config: Partial<RequestConfig<UpdateUserMutationRequest>> = {},
 ) {
-  const res = await client<UpdateUserMutationResponse, unknown, UpdateUserMutationRequest>({
+  const res = await client<UpdateUserMutationResponse, Error, UpdateUserMutationRequest>({
     method: 'put',
     url: `/user/${username}`,
     baseURL: 'https://petstore3.swagger.io/api/v3',