+++ examples/swr/src/gen/hooks/useCreateUser.ts
@@ -11,7 +11,7 @@ import type { SWRMutationConfiguration } from 'swr/mutation'
  * @link /user
  */
 async function createUser(data?: CreateUserMutationRequest, config: Partial<RequestConfig<CreateUserMutationRequest>> = {}) {
-  const res = await client<CreateUserMutationResponse, unknown, CreateUserMutationRequest>({
+  const res = await client<CreateUserMutationResponse, Error, CreateUserMutationRequest>({
     method: 'post',
     url: '/user',
     baseURL: 'https://petstore3.swagger.io/api/v3',