+++ examples/swr/src/gen/hooks/useLogoutUser.ts
@@ -9,7 +9,7 @@ import type { Key, SWRConfiguration } from 'swr'
  * @link /user/logout
  */
 async function logoutUser(config: Partial<RequestConfig> = {}) {
-  const res = await client<LogoutUserQueryResponse, unknown, unknown>({
+  const res = await client<LogoutUserQueryResponse, Error, unknown>({
     method: 'get',
     url: '/user/logout',
     baseURL: 'https://petstore3.swagger.io/api/v3',